+++ jwp-basic-gradle/src/main/java/next/controller/user/ListUserController.java
@@ -1,19 +1,23 @@
-package next.controller;
+package next.controller.user;
 
 import core.web.ModelAndView;
+import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import core.web.View;
+import next.controller.AbstractController;
+import next.dao.UserDao;
 import next.model.User;
 import next.view.JspView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class UpdateUserFormController extends AbstractController {
-    private static final long serialVersionUID = 1L;
-    private static final Logger log = LoggerFactory.getLogger(UpdateUserFormController.class);
+import java.sql.SQLException;
+
+public class ListUserController extends AbstractController {
+    private static final Logger log = LoggerFactory.getLogger(ListUserController.class);
+    private final UserDao userDao = new UserDao();
 
     @Override
     protected ModelAndView doGet(HttpServletRequest request, HttpServletResponse response) {
@@ -23,13 +27,14 @@ public class UpdateUserFormController extends AbstractController {
             log.debug("Only logged-in users can access.");
             return new ModelAndView(new JspView("redirect:/user/login.jsp"));
         }
-        User user = (User)value;
-        if (!user.isSameUser(request.getParameter("userId"))) {
-            log.debug("자신의 계정에만 접근할 수 있습니다.");
-            return new ModelAndView(new JspView("redirect:/user/list"));
-        }
 
-        request.setAttribute("user", user);
-        return new ModelAndView(new JspView("/user/update.jsp")).addModel("user", user);
+        List<User> users = null;
+        try {
+            users = userDao.findAll();
+            request.setAttribute("users", users);
+        } catch (SQLException e) {
+            log.error(e.toString());
+        }
+        return new ModelAndView(new JspView("/user/list.jsp")).addModel("users", users);
     }
 }
