+++ jwp-basic-gradle/src/main/java/next/controller/user/UpdateUserController.java
@@ -1,11 +1,11 @@
-package next.controller;
+package next.controller.user;
 
 import core.web.ModelAndView;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import core.web.View;
+import next.controller.AbstractController;
 import next.dao.UserDao;
 import next.model.User;
 import next.view.JspView;
