+++ jwp-basic-gradle/src/main/java/next/controller/user/UpdateUserFormController.java
@@ -1,24 +1,19 @@
-package next.controller;
+package next.controller.user;
 
-import core.db.DataBase;
 import core.web.ModelAndView;
-import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import core.web.View;
-import next.dao.UserDao;
+import next.controller.AbstractController;
 import next.model.User;
 import next.view.JspView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.sql.SQLException;
-
-public class ListUserController extends AbstractController {
-    private static final Logger log = LoggerFactory.getLogger(ListUserController.class);
-    private final UserDao userDao = new UserDao();
+public class UpdateUserFormController extends AbstractController {
+    private static final long serialVersionUID = 1L;
+    private static final Logger log = LoggerFactory.getLogger(UpdateUserFormController.class);
 
     @Override
     protected ModelAndView doGet(HttpServletRequest request, HttpServletResponse response) {
@@ -28,14 +23,13 @@ public class ListUserController extends AbstractController {
             log.debug("Only logged-in users can access.");
             return new ModelAndView(new JspView("redirect:/user/login.jsp"));
         }
-
-        List<User> users = null;
-        try {
-            users = userDao.findAll();
-            request.setAttribute("users", users);
-        } catch (SQLException e) {
-            log.error(e.toString());
+        User user = (User)value;
+        if (!user.isSameUser(request.getParameter("userId"))) {
+            log.debug("자신의 계정에만 접근할 수 있습니다.");
+            return new ModelAndView(new JspView("redirect:/user/list"));
         }
-        return new ModelAndView(new JspView("/user/list.jsp")).addModel("users", users);
+
+        request.setAttribute("user", user);
+        return new ModelAndView(new JspView("/user/update.jsp")).addModel("user", user);
     }
 }
