+++ jwp-basic-gradle/src/main/java/next/view/JsonView.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.PrintWriter;
-import next.controller.AnswerController;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
